Login, logout, modelUser, passato token Fatto il login con localstorege, passato il permesso a tutte le crus per pote accedere nelle pagine se loggat. Fatto logout con reidirizazione sul logincreato model del user. Passato il token valido alle crud per poter essere loggato su ogni pagina che voglio.

+++ src/app/servizi/firebase.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
+import { AuthService } from './../auth/auth.service';
 
 @Injectable({
   providedIn: 'root'
@@ -8,27 +9,27 @@ import { Observable } from 'rxjs';
 export class FirebaseService {
   percorso = 'https://corso-angular-3715a-default-rtdb.europe-west1.firebasedatabase.app/persone'
 
-  constructor(private httpClient: HttpClient) { }
+  constructor(private httpClient: HttpClient, private authService: AuthService) { }
 
 
-  inserPersona(body:{}): Observable<any> {
-    return this.httpClient.post(`${this.percorso}.json`, body);
+  inserPersona(body: {}): Observable<any> {
+    return this.httpClient.post(`${this.percorso}.json?auth=${this.authService.user.token}`, body);
   }
 
   getPersone() {
-    return this.httpClient.get(`${this.percorso}.json`)
+    return this.httpClient.get(`${this.percorso}.json?auth=${this.authService.user.token}`)
   }
 
   getPersona(id: string): Observable<any> {
-    return this.httpClient.get(`${this.percorso}/${id}.json`);
+    return this.httpClient.get(`${this.percorso}/${id}.json?auth=${this.authService.user.token}`);
   }
 
   deletePersona(id: string) {
-    return this.httpClient.delete(`${this.percorso}/${id}.json`);
+    return this.httpClient.delete(`${this.percorso}/${id}.json?auth=${this.authService.user.token}`);
   }
 
   patchPersona(id:string, body: {}) {
-    return this.httpClient.patch(`${this.percorso}/${id}.json`, body);
+    return this.httpClient.patch(`${this.percorso}/${id}.json?auth=${this.authService.user.token}`, body);
   }
   
 }
